rollback to float

+++ app/model/internal/UserDataOrder.scala
@@ -9,5 +9,5 @@ case class UserDataOrder(id: String,
                          contactInfo: ContactInfo,
                          photoUrl: String,
                          oneSignalId: String,
-                         scoring: Option[Double],
+                         scoring: Option[Float],
                          role: Option[String])
